argon2-cffi-bindings/pyproject.toml
[project]

name = "argon2_cffi_bindings_app"
version = "1.0.0"
description = ""
authors = [{name = "Flet Developer", email = "your@email.com"}]
dependencies = ["flet==0.25.0.dev3526", "argon2-cffi", "pytest"]

[tool.flet]
android.adaptive_icon_background = "#ffffff"